+++ gaios/Models/AppSettings.swift
@@ -5,7 +5,8 @@ struct AppSettings: Codable {
     static var shared = AppSettings()
     static let testnetIsVisible = "testnet_is_visible"
     static let experimental = "experimental"
-
+    static let rememberHWIsOff = "remember_HW_is_off"
+    
     var testnet: Bool {
         get { UserDefaults.standard.bool(forKey: AppSettings.testnetIsVisible) == true }
         set { UserDefaults.standard.set(newValue, forKey: AppSettings.testnetIsVisible) }
@@ -16,6 +17,11 @@ struct AppSettings: Codable {
         set { UserDefaults.standard.set(newValue, forKey: AppSettings.experimental) }
     }
 
+    var rememberHWIsOff: Bool {
+        get { UserDefaults.standard.bool(forKey: AppSettings.rememberHWIsOff) == true }
+        set { UserDefaults.standard.set(newValue, forKey: AppSettings.rememberHWIsOff) }
+    }
+    
     var gdkSettings: GdkSettings? {
         get { GdkSettings.read() }
         set { newValue?.write() }